requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
Flask==1.1.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4.1
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.9
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
